+++ crates/dbsp/examples/tutorial/tutorial7.rs
@@ -27,6 +27,7 @@ use size_of::SizeOf;
     IsNone,
 )]
 #[archive_attr(derive(Ord, Eq, PartialEq, PartialOrd))]
+#[archive(compare(PartialEq))]
 struct Record {
     location: String,
     date: i32,
@@ -50,7 +51,7 @@ struct Record {
     IsNone,
 )]
 #[archive_attr(derive(Ord, Eq, PartialEq, PartialOrd))]
-#[archive(compare(PartialEq, PartialOrd))]
+#[archive(compare(PartialEq))]
 struct VaxMonthly {
     count: u64,
     year: i32,
@@ -79,7 +80,7 @@ fn build_circuit(
                 r.daily_vaccinations.unwrap_or(0),
             )
         })
-        .aggregate_linear(|v| *v as isize);
+        .aggregate_linear(|v| *v as i64);
     let most_vax = monthly_totals
         .map_index(|(Tup3(l, y, m), sum)| {
             (
